move setWorkposThread - only used by embl_robot
 * having to import robot_lib from embl_robot caused
   a circular import, so move this function.
